refactor(5): extraer setup wizard a jobhunter/cli/setup.py
- _setup_screen, _ask, _ask_secret, _mask_secret, cmd_setup + 10 steps
- Preserva: nonlocal profile, cierre sobre cfg/TOTAL, state machine back
- job.py re-exporta los helpers publicos y cmd_setup
- tests/test_cli_setup.py: mask short/long/empty, _ask strip/back,
  _ask_secret enter mantiene, preview inline, sin current sin decoracion
